% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price.R
\name{get_price}
\alias{get_price}
\title{Get Stock Price Data}
\usage{
get_price(tickers, start_date, end_date, omit_na = FALSE, frequency = "daily")
}
\arguments{
\item{tickers}{A character vector or a concatenated string specifying one or several asset symbols.}

\item{start_date}{A string in the format "YYYY-MM-DD" indicating the start date of the data retrieval.}

\item{end_date}{A string in the format "YYYY-MM-DD" indicating the end date of the data retrieval.}

\item{omit_na}{A logical value indicating whether or not to omit missing observations. Set to TRUE to exclude any missing values from the retrieved data, or set to FALSE to include them.}

\item{frequency}{A character argument specifying the frequency of the data retrieval. Accepts the values "daily", "monthly", or "yearly" to indicate the desired frequency of the retrieved data.}
}
\value{
Returns a data.frame() containing historical closing prices for the provided asset symbol(s).
}
\description{
The function extends on the `getSymbols` function from the `quantmod` package,
developed by Jeffrey A. Ryan, to retrieve historical closing prices from Yahoo Finance.
The function provides the flexibility to retrieve historical price data for multiple assets.
It allows you to choose between retrieving daily data or aggregating the data into monthly or yearly averages.
}
\examples{

# Return daily historical closing prices for one asset ticker/symbol
get_price(
  tickers = "AAPL",
  start_date = "2020-01-01",
  end_date = "2020-05-01",
  omit_na = TRUE,
  frequency = "daily"
            )

# Return monthly historical closing prices for three asset tickers/symbols
get_price(
  tickers = c("AAPL", "MSFT", "TSLA"),
  start_date = "2020-01-01",
  end_date = "2020-05-01",
  omit_na = TRUE,
  frequency = "monthly"
            )

}
\author{
Christian André Berntsen
}
