% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{returns}
\alias{returns}
\title{Calculate Log Or Simple Returns}
\usage{
returns(get_price, type = c("log", "simple"))
}
\arguments{
\item{type}{Type of returns to calculate: "log" for logarithmic returns and "simple" for simple returns.}

\item{stock_price}{A data.frame() containing dates in the first column and stock prices in the remaining columns.}
}
\value{
Returns a data.frame() containing historical closing return observations for the provided data.frame()
}
\description{
This function calculates log or simple returns for a data frame. It is designed to be used with the `get_price()` function from the `assetr` package, but can also be applied to any data frame that has a date column in the first column and prices in the remaining columns.
}
\examples{

# Obtaining historical price data
prices <- get_price(c("AAPL", "TSLA"),
                    "2017-01-01",
                    "2023-01-01",
                    omit_na = T,
                    frequency = "daily")

# Calculate log returns
returns(prices, type = "log")

}
